+++ app/modules/admin/Module.php
@@ -11,7 +11,6 @@ class Module extends \yii\base\Module
 	 */
 	public $layout = 'main';
 
-
 	/**
 	 * @inheritdoc
 	 */