+++ app/modules/admin/assets/AssetBundle.php
@@ -2,20 +2,20 @@
 
 namespace app\modules\admin\assets;
 
+use app\assets\FontAwesomeAsset;
+
 class AssetBundle extends \yii\web\AssetBundle
 {
 	public $sourcePath = '@app/modules/admin/assets';
 
 	public $css = [
-		'css/extra.css'
+		'css/admin.css'
 	];
 
 	public $js = [];
 
 	public $depends = [
-		'dmstr\web\AdminLteAsset',
-		//'app\modules\admin\assets\AdminltePluginsAssetBundle',
-		'yii\web\YiiAsset',
-		'yii\bootstrap\BootstrapAsset',
+		AdminLteAsset::class,
+		FontAwesomeAsset::class,
 	];
 }